// src/types/settings.ts
export interface StoreSettings {
  storeName: string;
  logoUrl: string;
  whatsappNumber: string;
  deliveryFee: number;
  isFreeDelivery: boolean;
  deliveryEnabled: boolean;
  pickupEnabled: boolean;
  storeAddress: string;
  openingHours: string;
  instagramUrl: string;
  welcomeMessage: string;
}